chore: refactor withdraw & minor lib file locations (#34)
* feat. add persistent store

* wip: fix persist networks

* feat: persist all networks

* refactor: create evm toolkit lib module & cleanup code

* wip: refactor

* fix: typo

* chore: refactor files
