--- tools/evm-toolkit/src/components/withdraw/iscmagic/index.ts
@@ -1,2 +1,0 @@
-export * from './iscmagic';
-export * from './types/web3-v1-contracts';
