+++ tools/evm-toolkit/src/lib/assets/multicall3.json
