--- tools/evm-toolkit/src/lib/balance.ts
@@ -1,1 +1,0 @@
-export const Mi: number = 1e6;
