+++ tools/evm-toolkit/src/lib/common/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './components.interfaces'
