--- tools/evm-toolkit/src/lib/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from 'components.interfaces'
