+++ tools/evm-toolkit/src/lib/iscmagic/index.ts
@@ -1,0 +1,3 @@
+export * from './iscmagic';
+export * from './types/web3-v1-contracts';
+export * from './utils';
