+++ tools/evm-toolkit/src/lib/native-token/constants.ts
@@ -1,0 +1,3 @@
+import type { INativeTokenMetaDataCacheMap } from './interfaces';
+
+export const nativeTokenMetadataCache: INativeTokenMetaDataCacheMap = {};
