+++ tools/evm-toolkit/src/lib/native-token/index.ts
@@ -1,5 +1,3 @@
 export * from './actions'
 export * from './constants'
 export * from './interfaces'
-export * from './stores'
-export * from './utils'
