+++ tools/evm-toolkit/src/lib/simple-buffer-cursor.ts
