+++ tools/evm-toolkit/src/lib/withdraw/constants.ts
@@ -1,5 +1,5 @@
-import iscAbiAsText from '../../assets/ISCSandbox.abi?raw';
-import multiCallAbiAsText from '../../assets/multicall3.json?raw';
+import multiCallAbiAsText from '$lib/assets/multicall3.json?raw';
+import iscAbiAsText from '$lib/assets/ISCSandbox.abi?raw';
 
 export const gasFee = 21000;
 export const iscAbi = JSON.parse(iscAbiAsText);
