+++ tools/evm-toolkit/src/lib/withdraw/utils.ts
@@ -1,7 +1,11 @@
-import { Converter } from '@iota/util.js';
-import { waspAddrBinaryFromBech32 } from './../../lib/bech32';
 import type { SingleNodeClient } from '@iota/iota.js';
-import type { INativeToken } from '../../lib/native_token';
+import {
+  Bech32Helper,
+  type IEd25519Address
+} from '@iota/iota.js';
+import { Converter } from '@iota/util.js';
+
+import type { INativeToken } from '$lib/native-token';
 
 export function getBalanceParameters(agentID: Uint8Array) {
   return {
@@ -12,17 +16,6 @@ export function getBalanceParameters(agentID: Uint8Array) {
       },
     ],
   };
-}
-
-export interface INativeTokenWithdraw {
-  /**
-   * Identifier of the native token.
-   */
-  ID: string;
-  /**
-   * Amount of native tokens of the given Token ID.
-   */
-  amount: bigint;
 }
 
 export async function withdrawParameters(
@@ -92,4 +85,29 @@ export async function withdrawParameters(
   ];
   console.log(parameters)
   return parameters;
+}
+
+export async function waspAddrBinaryFromBech32(
+  nodeClient: SingleNodeClient,
+  bech32String: string,
+) {
+  const protocolInfo = await nodeClient.info();
+
+  const receiverAddr = Bech32Helper.addressFromBech32(
+    bech32String,
+    protocolInfo.protocol.bech32Hrp,
+  );
+
+  const address: IEd25519Address = receiverAddr as IEd25519Address;
+
+  const receiverAddrBinary = Converter.hexToBytes(address.pubKeyHash);
+  //  // AddressEd25519 denotes an Ed25519 address.
+  // AddressEd25519 AddressType = 0
+  // // AddressAlias denotes an Alias address.
+  // AddressAlias AddressType = 8
+  // // AddressNFT denotes an NFT address.
+  // AddressNFT AddressType = 16
+  //
+  // 0 is the ed25519 prefix
+  return new Uint8Array([0, ...receiverAddrBinary]);
 }
