+++ tools/evm-toolkit/src/routes/sections/index.ts
@@ -1,1 +1,2 @@
 export { default as FaucetSection } from './faucet-section.svelte';
+export { default as WithdrawSection } from './withdraw-section.svelte';
